unit test attempt
So I've tried to write tests in between the many things that happened this past week but I ended up erasing and rewriting various tests in a directionless manner. This was one of the many tests I attempted to write (keep in mind, this is my first time hearing about/being exposed to/writing tests).

This test failed because I didn't define var AlertView in the CrashViewController.swift. I didn't want to mess with the view controller until I was sure that the test was alright to avoid as many merge conflicts as possible.

I intend for this unit test to verify that the alert view controller shows up, or that the presentCrashAlert function works like it's supposed to.